# docs/api.yaml
openapi: 3.0.2
info:
  version: "0.1.0"
  title: 'Wallet Permissions service'
  description: Permissions API for developers.
security:
  - bearerAuth: []
paths:
  '/permissions/private/v1/admin/group':
    get:
      tags:
        - Group
      summary: Shows a list of admin groups.
      description: |
        Available for users with "view_settings" permission.
        \
        Returns all groups if "userId" is not passed. Otherwise returns groups for user by passed id.
      operationId: getGroups
      parameters:
        - name: userId
          in: query
          description: Filter by user id (get group for specific user).
          required: false
          schema:
            type: string
        - name: scope
          in: query
          description: Filter by scope.
          required: false
          schema:
            type: string
            enum: ["admin", "client"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      object:
                        type: string
                        enum:
                          - list
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Group'
        '403':
          description: forbidden
    post:
      tags:
        - Group
      summary: Creates a new admin group.
      description: |
        Available for admins with "create_settings" permission.
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroup'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Group'
        '400':
          description: group name is not specified or duplicate record found
        '403':
          description: forbidden
  '/permissions/private/v1/admin/group/{id}':
    post:
      tags:
        - Group
      summary: Updates an admin group.
      description: Available for admins with "modify_settings" permission.
      operationId: updateGroup
      parameters:
        - name: id
          in: path
          description: An user group id.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroup'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Group'
        '400':
          description: duplicate record found
        '403':
          description: forbidden
    get:
      tags:
        - Group
      summary: Shows a group by id.
      operationId: getGroupById
      parameters:
        - name: id
          in: path
          description: An user group id.
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Group'
        '403':
          description: forbidden

    delete:
      tags:
        - Group
      summary: Deletes a group.
      operationId: deleteGroupById
      parameters:
        - name: id
          in: path
          description: An user group id.
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: successfully deleted
        '403':
          description: forbidden
        '409':
          description: group cannot be deleted as long as administrators are present in the group
...
components:
  schemas:
    CreateGroup:
      type: object
      properties:
        name:
          type: string
          description: Group name (required)
        description:
          type: string
          description: Optional group description
        scope:
          type: string
          description: Group scope (required)
          enum:
            - client
            - admin
        actions:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              enabled:
                type: boolean
      required:
        - name
        - scope
    UpdateGroup:
      type: object
      properties:
        name:
          type: string
          description: New group name (optional)
        description:
          type: string
          description: Add or edit group description (optional)
        actions:
          description: Enable/disable actions for the group
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              enabled:
                type: boolean
...
    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        object:
          type: string
          description: object class
          enum:
            - group
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT